requirements.txt
torch
opencv-python
onnxruntime-gpu[cuda,cudnn]
numpy
supervision
yt-dlp
trackers